Feat: Add AppConfig for security configuration and update AuthController for cleaner code
