+++ Backend/gestao-projetos/src/main/java/com/gestaoprojetos/config/AppConfig.java
@@ -1,0 +1,31 @@
+package com.gestaoprojetos.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
+
+@Configuration
+public class AppConfig {
+
+    /**
+     * Bean de PasswordEncoder (BCrypt) separado.
+     * Agora toda injeção de PasswordEncoder virá desta classe.
+     */
+    @Bean
+    public PasswordEncoder passwordEncoder() {
+        return new BCryptPasswordEncoder();
+    }
+
+    /**
+     * Expõe AuthenticationManager para uso externo (no AuthController, etc.).
+     */
+    @Bean
+    public AuthenticationManager authenticationManager(
+            AuthenticationConfiguration authConfig
+    ) throws Exception {
+        return authConfig.getAuthenticationManager();
+    }
+}
